refactor(vue-engine): extract vueDeleteFile from VolarEngine
deleteFile coordinates two distinct concerns (TS import removal + Vue
import removal). Extracting it to delete-file.ts mirrors the pattern
ts-engine/delete-file.ts already uses and leaves VolarEngine as a thin
dispatcher.

Clarifies handoff.md chore definition: entries are for deferred admin
tasks; inline refactors spotted during a session don't need an entry.
